Moves layers around in `network/*`
